# thf-conference-api/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "THF conference application API.\n"
  version: "1.0.0"
  title: "THF Conference App"
  termsOfService: ""
  contact:
    url: "https://thf.totvs.com.br/"
host: "thfservices.totvs.com.br"
basePath: "/conference-api/api/v1"
schemes:
- "http"
consumes:
- "application/json"
produces:
- "application/json"
paths:
  /conferences/:
    get:
      tags:
      - "Conferences"
      summary: "Load the list of conferences"
      operationId: "conferencesGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Status 200 - OK"
          schema:
            $ref: "#/definitions/conferencesResponse"
          examples:
            application/json: "{\n\t\"hasNext\": false,\n\t\"items\": [{\n\t\t\"id\"\
              : 1,\n\t\t\"title\": \"THF conference 2017\",\n\t\t\"date\": \"2017-07-11T00:00:00Z\"\
              ,\n\t\t\"location\": \"Av. Santos Dumont, 831 - Santo Antônio, Joinville\
              \ - SC\",\n\t\t\"description\": \"Conference organized by THF\"\n\t\
              }, {\n\t\t\"id\": 2,\n\t\t\"title\": \"THF conference 2018\",\n\t\t\"\
              date\": \"2018-08-11T00:00:00Z\",\n\t\t\"location\": \"Av. Santos Dumont,\
              \ 831 - Santo Antônio, Joinville - SC\",\n\t\t\"description\": \"Conference\
              \ organized by THF\"\n\t}]\n}"
        500:
          description: "Status 500 - Internal Server Error"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"500\",\n\t\"message\": \"Internal\
              \ Server Error\",\n\t\"detailedMessage\": \"High system resource usage\"\
              \n}"
      x-swagger-router-controller: "Conferences"
    x-restlet:
      section: "Conference"
  /conferences/{id}:
    get:
      tags:
      - "Conferences"
      summary: "Load an individual conference"
      operationId: "conferencesIdGET"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Conference id."
        required: true
        type: "string"
        x-example: "1"
      responses:
        200:
          description: "Status 200 - OK"
          schema:
            $ref: "#/definitions/conference"
          examples:
            application/json: "{\n\t\"id\": 2,\n\t\"title\": \"THF conference 2018\"\
              ,\n\t\"date\": \"2018-08-11T00:00:00Z\",\n\t\"location\": \"Av. Santos\
              \ Dumont, 831 - Santo Antônio, Joinville - SC\",\n\t\"description\"\
              : \"Conference organized by THF\"\n}"
        404:
          description: "Status 404 - Not Found"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"404\",\n\t\"message\": \"Not Found\"\
              ,\n\t\"detailedMessage\": \"Conference not found\"\n}"
        500:
          description: "Status 500 - Internal Server Error"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"500\",\n\t\"message\": \"Internal\
              \ Server Error\",\n\t\"detailedMessage\": \"High system resource usage\"\
              \n}"
      x-swagger-router-controller: "Conferences"
    put:
      tags:
      - "Conferences"
      summary: "Update a conference"
      operationId: "conferencesIdPUT"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Conference id."
        required: true
        type: "string"
        x-example: "1"
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/conference"
      responses:
        200:
          description: "Status 200 - OK"
          schema:
            $ref: "#/definitions/conference"
          examples:
            application/json: "{\n\t\"id\": 2,\n\t\"title\": \"THF conference 2018\"\
              ,\n\t\"date\": \"2018-08-11T00:00:00Z\",\n\t\"location\": \"Av. Santos\
              \ Dumont, 831 - Santo Antônio, Joinville - SC\",\n\t\"description\"\
              : \"Conference organized by THF\"\n}"
        404:
          description: "Status 404 - Not Found"
          schema:
            $ref: "#/definitions/errorResponse"
        500:
          description: "Status 500 - Internal Server Error"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"500\",\n\t\"message\": \"Internal\
              \ Server Error\",\n\t\"detailedMessage\": \"High system resource usage\"\
              \n}"
      x-swagger-router-controller: "Conferences"
    x-restlet:
      section: "Conference"
  /lectures/:
    get:
      tags:
      - "Lectures"
      summary: "Load the list of lectures"
      operationId: "lecturesGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Status 200"
          schema:
            $ref: "#/definitions/lecturesResponse"
          examples:
            application/json: "{\n\t\"hasNext\": false,\n\t\"items\": [{\n\t\t\"id\"\
              : \"1\",\n\t\t\"title\": \"Unit testing for beginners\",\n\t\t\"room\"\
              : \"2B\",\n\t\t\"startTime\": \"08:30:00\",\n\t\t\"endTime\": \"08:45:00\"\
              ,\n\t\t\"description\": \"Unit testing for beginners\",\n\t\t\"trackId\"\
              : \"4\",\n\t\t\"speakerId\": \"4\"\n\t}, {\n\t\t\"id\": \"2\",\n\t\t\
              \"title\": \"Chrome devtools\",\n\t\t\"room\": \"2C\",\n\t\t\"startTime\"\
              : \"08:45:00\",\n\t\t\"endTime\": \"09:00:00\",\n\t\t\"description\"\
              : \"Chrome devtools\",\n\t\t\"trackId\": \"6\",\n\t\t\"speakerId\":\
              \ \"3\"\n\t}]\n}"
        500:
          description: "Status 500 - Internal Server Error"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"500\",\n\t\"message\": \"Internal\
              \ Server Error\",\n\t\"detailedMessage\": \"High system resource usage\"\
              \n}"
      x-swagger-router-controller: "Lectures"
    post:
      tags:
      - "Lectures"
      summary: "Create a new lecture"
      operationId: "lecturesPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/lecture"
      responses:
        200:
          description: "Status 200 - OK"
          schema:
            $ref: "#/definitions/lecture"
        500:
          description: "Status 500 - Internal Server Error"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"500\",\n\t\"message\": \"Internal\
              \ Server Error\",\n\t\"detailedMessage\": \"High system resource usage\"\
              \n}"
      x-swagger-router-controller: "Lectures"
    x-restlet:
      section: "Lectures"
  /lectures/{id}:
    get:
      tags:
      - "Lectures"
      summary: "Load an individual lecture"
      operationId: "lecturesIdGET"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Lecture id."
        required: true
        type: "number"
        x-example: 1
      responses:
        200:
          description: "Status 200 - OK"
          schema:
            $ref: "#/definitions/lecture"
          examples:
            application/json: "{\n\t\t\"id\": \"1\",\n\t\t\"title\": \"Unit testing\
              \ for beginners\",\n\t\t\"room\": \"2B\",\n\t\t\"startTime\": \"08:30:00\"\
              ,\n\t\t\"endTime\": \"08:45:00\",\n\t\t\"description\": \"Unit testing\
              \ for beginners\",\n\t\t\"trackId\": \"4\",\n\t\t\"speakerId\": \"4\"\
              \n}"
        404:
          description: "Status 404 - Not Found"
          schema:
            $ref: "#/definitions/errorResponse"
        500:
          description: "Status 500 - Internal Server Error"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"500\",\n\t\"message\": \"Internal\
              \ Server Error\",\n  \t\"detailedMessage\": \"High system resource usage\"\
              \n}"
      x-swagger-router-controller: "Lectures"
    put:
      tags:
      - "Lectures"
      summary: "Update a lecture"
      operationId: "lecturesIdPUT"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Lecture id."
        required: true
        type: "number"
        x-example: 1
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/lecture"
      responses:
        200:
          description: "Status 200 - OK"
          schema:
            $ref: "#/definitions/lecture"
          examples:
            application/json: "{\n\t\t\"id\": \"1\",\n\t\t\"title\": \"Unit testing\
              \ for beginners\",\n\t\t\"room\": \"2B\",\n\t\t\"startTime\": \"08:30:00\"\
              ,\n\t\t\"endTime\": \"08:45:00\",\n\t\t\"description\": \"Unit testing\
              \ for beginners\",\n\t\t\"trackId\": \"4\",\n\t\t\"speakerId\": \"4\"\
              \n}"
        404:
          description: "Status 404 - Not Found"
          schema:
            $ref: "#/definitions/errorResponse"
        500:
          description: "Status 500 - Internal Server Error"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"500\",\n\t\"message\": \"Internal\
              \ Server Error\",\n  \t\"detailedMessage\": \"High system resource usage\"\
              \n}"
      x-swagger-router-controller: "Lectures"
    delete:
      tags:
      - "Lectures"
      summary: "Delete a lecture"
      operationId: "lecturesIdDELETE"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Lecture id."
        required: true
        type: "number"
        x-example: 1
      responses:
        204:
          description: "Status 204 - No Content"
        404:
          description: "Status 404 - Not Found"
          schema:
            $ref: "#/definitions/errorResponse"
        500:
          description: "Status 500 - Internal Server Error"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"500\",\n\t\"message\": \"Internal\
              \ Server Error\",\n  \t\"detailedMessage\": \"High system resource usage\"\
              \n}"
      x-swagger-router-controller: "Lectures"
    x-restlet:
      section: "Lectures"
  /notes/:
    get:
      tags:
      - "Notes"
      summary: "Load the list of notes"
      operationId: "notesGET"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "Status 200"
          schema:
            $ref: "#/definitions/notesResponse"
          examples:
            application/json: "{\n\t\"hasNext\": false,\n\t\"items\": [{\n\t\t\"id\"\
              : \"1\",\n\t\t\"title\": \"Chrome DevTools\",\n\t\t\"text\": \"Note\
              \ of Chrome DevTools\",\n\t\t\"lectureId\": \"1\",\n\t\t\"userId\":\
              \ \"2\"\n\t}, {\n\t\t\"id\": \"2\",\n\t\t\"title\": \"Unit testing for\
              \ beginners\",\n\t\t\"text\": \"Note of unit testing for beginners\"\
              ,\n\t\t\"lectureId\": \"2\",\n\t\t\"userId\": \"1\"\n\t}]\n}"
        500:
          description: "Status 500 - Internal Server Error"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"500\",\n\t\"message\": \"Internal\
              \ Server Error\",\n  \t\"detailedMessage\": \"High system resource usage\"\
              \n}"
      x-swagger-router-controller: "Notes"
    post:
      tags:
      - "Notes"
      summary: "Create a new note"
      operationId: "notesPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/note"
      responses:
        200:
          description: "Status 200 - OK"
          schema:
            $ref: "#/definitions/note"
        500:
          description: "Status 500 - Internal Server Error"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"500\",\n\t\"message\": \"Internal\
              \ Server Error\",\n  \t\"detailedMessage\": \"High system resource usage\"\
              \n}"
      x-swagger-router-controller: "Notes"
    x-restlet:
      section: "Notes"
  /notes/{id}:
    get:
      tags:
      - "Notes"
      summary: "Load an individual note"
      operationId: "notesIdGET"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Note id."
        required: true
        type: "number"
        x-example: 1
      responses:
        200:
          description: "Status 200 - OK"
          schema:
            $ref: "#/definitions/note"
          examples:
            application/json: "{\n\t\t\"id\": \"1\",\n\t\t\"title\": \"Chrome DevTools\"\
              ,\n\t\t\"text\": \"Note of Chrome DevTools\",\n\t\t\"lectureId\": \"\
              1\",\n\t\t\"userId\": \"2\"\n}"
        404:
          description: "Status 404 - Not Found"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"404\",\n\t\"message\": \"Not Found\"\
              ,\n\t\"detailedMessage\": \"Note not found\"\n}"
        500:
          description: "Status 500 - Internal Server Error"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"500\",\n\t\"message\": \"Internal\
              \ Server Error\",\n  \t\"detailedMessage\": \"High system resource usage\"\
              \n}"
      x-swagger-router-controller: "Notes"
    put:
      tags:
      - "Notes"
      summary: "Update a note"
      operationId: "notesIdPUT"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Note id."
        required: true
        type: "number"
        x-example: 1
      - in: "body"
        name: "body"
        required: true
        schema:
          $ref: "#/definitions/note"
      responses:
        200:
          description: "Status 200 - OK"
          schema:
            $ref: "#/definitions/note"
        404:
          description: "Status 404 - Not Found"
          schema:
            $ref: "#/definitions/errorResponse"
        500:
          description: "Status 500 - Internal Server Error"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"500\",\n\t\"message\": \"Internal\
              \ Server Error\",\n  \t\"detailedMessage\": \"High system resource usage\"\
              \n}"
      x-swagger-router-controller: "Notes"
    delete:
      tags:
      - "Notes"
      summary: "Delete a note"
      operationId: "notesIdDELETE"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "Note id."
        required: true
        type: "number"
        x-example: 1
      responses:
        204:
          description: "Status 204 - No Content"
        404:
          description: "Status 404 - Not Found"
          schema:
            $ref: "#/definitions/errorResponse"
        500:
          description: "Status 500 - Internal Server Error"
          schema:
            $ref: "#/definitions/errorResponse"
          examples:
            application/json: "{\n\t\"code\": \"500\",\n\t\"message\": \"Internal\
              \ Server Error\",\n  \t\"detailedMessage\": \"High system resource usage\"\
              \n}"
      x-swagger-router-controller: "Notes"
    x-restlet:
      section: "Notes"
...
definitions:
  errorResponse:
    type: "object"
    required:
    - "code"
    - "detailedMessage"
    - "message"
    properties:
      code:
        type: "string"
        description: "Error code."
      message:
        type: "string"
        description: "Error message."
      detailedMessage:
        type: "string"
        description: "Message detail."
      helpUrl:
        type: "string"
        description: "Help URL."
      details:
        type: "string"
        description: "Details."
    description: "This general error structure is used throughout this API."
    x-restlet:
      section: "General"
  conference:
    type: "object"
    required:
    - "date"
    - "description"
    - "location"
    - "title"
    properties:
      id:
        type: "string"
        description: "Conference id."
      title:
        type: "string"
        description: "Conference title."
      date:
        type: "string"
        format: "date-time"
        description: "Conference date."
      location:
        type: "string"
        description: "Conference location."
      description:
        type: "string"
        description: "Conference description."
    x-restlet:
      section: "Conference"
  conferencesResponse:
    type: "object"
    required:
    - "hasNext"
    - "items"
    properties:
      hasNext:
        type: "boolean"
        description: "Indicate if has next page of data."
      items:
        type: "array"
        description: "Conference list."
        items:
          $ref: "#/definitions/conference"
    x-restlet:
      section: "Conference"
  lecture:
    type: "object"
    required:
    - "description"
    - "endTime"
    - "room"
    - "speakerId"
    - "startTime"
    - "title"
    - "trackId"
    properties:
      id:
        type: "string"
        description: "Lecture id."
      title:
        type: "string"
        description: "Lecture title."
      description:
        type: "string"
        description: "Lecture description."
      room:
        type: "string"
        description: "Lecture room."
      startTime:
        type: "string"
        description: "Lecture start."
      endTime:
        type: "string"
        description: "Lecture end."
      trackId:
        type: "string"
        description: "Lecture track."
      speakerId:
        type: "string"
        description: "Speaker id."
    example:
      trackId: "trackId"
      description: "description"
      startTime: "startTime"
      speakerId: "speakerId"
      id: "id"
      endTime: "endTime"
      title: "title"
      room: "room"
    x-restlet:
      section: "Lectures"
...
  lecturesResponse:
    type: "object"
    required:
    - "hasNext"
    - "items"
    properties:
      hasNext:
        type: "boolean"
        description: "Indicate if has next page of data."
      items:
        type: "array"
        description: "Lecture list."
        items:
          $ref: "#/definitions/lecture"
    x-restlet:
      section: "Lectures"
  note:
    type: "object"
    required:
    - "lectureId"
    - "text"
    - "title"
    - "userId"
    properties:
      id:
        type: "string"
        description: "Note id."
      title:
        type: "string"
        description: "Note title."
      text:
        type: "string"
        description: "Note text."
      lectureId:
        type: "string"
        description: "Lecture id."
      userId:
        type: "string"
        description: "User id."
    example:
      id: "id"
      text: "text"
      title: "title"
      userId: "userId"
      lectureId: "lectureId"
    x-restlet:
      section: "Notes"
  notesResponse:
    type: "object"
    required:
    - "hasNext"
    - "items"
    properties:
      hasNext:
        type: "boolean"
        description: "Indicate if has next page of data."
      items:
        type: "array"
        description: "Note list."
        items:
          $ref: "#/definitions/note"
    x-restlet:
      section: "Notes"
...